+++ traffic_ops/traffic_ops_golang/deliveryservice/servers/servers.go
@@ -319,17 +319,19 @@ func GetReplaceHandler(w http.ResponseWriter, r *http.Request) {
 		api.HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
 		return
 	}
-	serverNames := []string{}
-	for _, s := range servers {
-		name, _, err := dbhelpers.GetServerNameFromID(inf.Tx.Tx, s)
-		if err != nil {
-			api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, err, nil)
-			return
-		}
-		serverNames = append(serverNames, name)
+	serverNamesAndTypes, err := dbhelpers.GetServerHostNamesAndTypesFromIDs(inf.Tx.Tx, servers)
+	if err != nil {
+		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, err, nil)
+		return
 	}
 
-	usrErr, sysErr, status := ValidateServerCapabilities(ds.ID, serverNames, inf.Tx.Tx)
+	userErr = ValidateDSSAssignments(ds, serverNamesAndTypes)
+	if userErr != nil {
+		api.HandleErr(w, r, inf.Tx.Tx, http.StatusBadRequest, userErr, nil)
+		return
+	}
+
+	usrErr, sysErr, status := ValidateServerCapabilities(ds.ID, serverNamesAndTypes, inf.Tx.Tx)
 	if usrErr != nil || sysErr != nil {
 		api.HandleErr(w, r, inf.Tx.Tx, status, usrErr, sysErr)
 		return
@@ -399,7 +401,19 @@ func GetCreateHandler(w http.ResponseWriter, r *http.Request) {
 	payload.XmlId = dsName
 	serverNames := payload.ServerNames
 
-	usrErr, sysErr, status := ValidateServerCapabilities(ds.ID, serverNames, inf.Tx.Tx)
+	serverNamesAndTypes, err := dbhelpers.GetServerTypesFromHostNames(inf.Tx.Tx, serverNames)
+	if err != nil {
+		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, err, nil)
+		return
+	}
+
+	userErr = ValidateDSSAssignments(ds, serverNamesAndTypes)
+	if userErr != nil {
+		api.HandleErr(w, r, inf.Tx.Tx, http.StatusBadRequest, userErr, nil)
+		return
+	}
+
+	usrErr, sysErr, status := ValidateServerCapabilities(ds.ID, serverNamesAndTypes, inf.Tx.Tx)
 	if usrErr != nil || sysErr != nil {
 		api.HandleErr(w, r, inf.Tx.Tx, status, usrErr, sysErr)
 		return
@@ -430,24 +444,26 @@ func GetCreateHandler(w http.ResponseWriter, r *http.Request) {
 	api.WriteResp(w, r, tc.DeliveryServiceServers{payload.ServerNames, payload.XmlId})
 }
 
-// ValidateServerCapabilities checks that the delivery service's requirements are met by each server to be assigned.
-func ValidateServerCapabilities(dsID int, serverNames []string, tx *sql.Tx) (error, error, int) {
-	nonOriginServerNames := []string{}
-	nonOriginTypeQuery := `
-	SELECT ARRAY(
-		SELECT s.host_name
-		FROM server s
-		JOIN type t ON s.type = t.id
-		WHERE t.name LIKE 'EDGE%' AND s.host_name = ANY($1)
-	)`
-
-	serverNamePqArray := pq.Array(serverNames)
-
-	if err := tx.QueryRow(nonOriginTypeQuery, serverNamePqArray).Scan(pq.Array(&nonOriginServerNames)); err != nil {
-		if err == sql.ErrNoRows {
-			return nil, nil, http.StatusOK
+// ValidateDSSAssignments returns an error if the given servers cannot be assigned to the given delivery service.
+func ValidateDSSAssignments(ds DSInfo, servers []dbhelpers.ServerHostNameAndType) error {
+	if ds.Topology == nil {
+		return nil
+	}
+	for _, s := range servers {
+		if s.Type != tc.OriginTypeName {
+			return errors.New("only servers of type ORG may be assigned to topology-based delivery services")
 		}
-		return nil, err, http.StatusInternalServerError
+	}
+	return nil
+}
+
+// ValidateServerCapabilities checks that the delivery service's requirements are met by each server to be assigned.
+func ValidateServerCapabilities(dsID int, serverNamesAndTypes []dbhelpers.ServerHostNameAndType, tx *sql.Tx) (error, error, int) {
+	nonOriginServerNames := []string{}
+	for _, s := range serverNamesAndTypes {
+		if strings.HasPrefix(s.Type, tc.EdgeTypePrefix) {
+			nonOriginServerNames = append(nonOriginServerNames, s.HostName)
+		}
 	}
 
 	var sCaps []string
@@ -665,6 +681,7 @@ type DSInfo struct {
 	SigningAlgorithm     *string
 	CacheURL             *string
 	MaxOriginConnections *int
+	Topology             *string
 }
 
 // GetDSInfo loads the DeliveryService fields needed by Delivery Service Servers from the database, from the ID. Returns the data, whether the delivery service was found, and any error.
@@ -678,7 +695,8 @@ SELECT
   ds.regex_remap,
   ds.signing_algorithm,
   ds.cacheurl,
-  ds.max_origin_connections
+  ds.max_origin_connections,
+  ds.topology
 FROM
   deliveryservice ds
   JOIN type tp ON ds.type = tp.id
@@ -686,7 +704,7 @@ WHERE
   ds.id = $1
 `
 	di := DSInfo{ID: id}
-	if err := tx.QueryRow(qry, id).Scan(&di.Name, &di.Type, &di.EdgeHeaderRewrite, &di.MidHeaderRewrite, &di.RegexRemap, &di.SigningAlgorithm, &di.CacheURL, &di.MaxOriginConnections); err != nil {
+	if err := tx.QueryRow(qry, id).Scan(&di.Name, &di.Type, &di.EdgeHeaderRewrite, &di.MidHeaderRewrite, &di.RegexRemap, &di.SigningAlgorithm, &di.CacheURL, &di.MaxOriginConnections, &di.Topology); err != nil {
 		if err == sql.ErrNoRows {
 			return DSInfo{}, false, nil
 		}
@@ -707,7 +725,8 @@ SELECT
   ds.regex_remap,
   ds.signing_algorithm,
   ds.cacheurl,
-  ds.max_origin_connections
+  ds.max_origin_connections,
+  ds.topology
 FROM
   deliveryservice ds
   JOIN type tp ON ds.type = tp.id
@@ -715,7 +734,7 @@ WHERE
   ds.xml_id = $1
 `
 	di := DSInfo{Name: dsName}
-	if err := tx.QueryRow(qry, dsName).Scan(&di.ID, &di.Type, &di.EdgeHeaderRewrite, &di.MidHeaderRewrite, &di.RegexRemap, &di.SigningAlgorithm, &di.CacheURL, &di.MaxOriginConnections); err != nil {
+	if err := tx.QueryRow(qry, dsName).Scan(&di.ID, &di.Type, &di.EdgeHeaderRewrite, &di.MidHeaderRewrite, &di.RegexRemap, &di.SigningAlgorithm, &di.CacheURL, &di.MaxOriginConnections, &di.Topology); err != nil {
 		if err == sql.ErrNoRows {
 			return DSInfo{}, false, nil
 		}